allow to skip YDB deploy, add --restart option to skip fomatting/bs init

+++ db_installers/ydb/setup.sh
@@ -7,7 +7,7 @@ log() {
 }
 
 usage() {
-    echo "Usage: setup.sh --ydbd <PATH_TO_YDBD_TAR> --ydbd-url <YDB_TAR_URL> --config <PATH_TO_SETUP_CONFIG> [--stop]"
+    echo "Usage: setup.sh [--ydbd <PATH_TO_YDBD_TAR>] [--ydbd-url <YDB_TAR_URL>] --config <PATH_TO_SETUP_CONFIG> [--stop] [--restart]"
 }
 
 if ! command -v parallel-ssh &> /dev/null
@@ -30,6 +30,8 @@ while [[ $# -gt 0 ]]; do case $1 in
         shift;;
     --stop)
         stop_ydb=1;;
+    --restart)
+        restart_ydb=1;;
     --help|-h)
         usage
         exit;;
@@ -37,12 +39,6 @@ while [[ $# -gt 0 ]]; do case $1 in
         usage
         exit;;
 esac; shift; done
-
-if [[ -z "$ydbd_tar" && -z "$ydbd_url" && -z $stop_ydb ]]; then
-    echo "ERROR: you must specify either option: --ydbd, --ydbd-url, --stop"
-    usage
-    exit 1
-fi
 
 if [[ -z "$setup_config" ]]; then
     echo "ERROR: Setup config is not specified"
@@ -77,7 +73,6 @@ if [[ $DYNNODE_COUNT -gt ${#DYNNODE_TASKSET_CPU[@]} ]]; then
     exit 1
 fi
 
-
 eval "HOSTS_FILE=$HOSTS_FILE"
 eval "CONFIG_DIR=$CONFIG_DIR"
 
@@ -111,46 +106,47 @@ if [ $stop_ydb -eq 1 ]; then
     exit 0
 fi
 
-log "Deploy"
+if [[ -n "$ydbd_tar" || -n "$ydbd_url" ]]; then
+    log "Deploy"
 
-$debug parallel-ssh -i -h "$HOSTS_FILE" -t 0 -p 20 "\
-    sudo rm -rf $YDB_SETUP_PATH;   \
-    sudo mkdir -p $YDB_SETUP_PATH; \
-    sudo chown $USER $YDB_SETUP_PATH; \
-    mkdir $YDB_SETUP_PATH/cfg $YDB_SETUP_PATH/logs"
+    $debug parallel-ssh -i -h "$HOSTS_FILE" -t 0 -p 20 "\
+        sudo rm -rf $YDB_SETUP_PATH;   \
+        sudo mkdir -p $YDB_SETUP_PATH; \
+        sudo chown $USER $YDB_SETUP_PATH; \
+        mkdir $YDB_SETUP_PATH/cfg $YDB_SETUP_PATH/logs"
 
-if [[ $? -ne 0 ]]; then
-    echo "ERROR: can't create $YDB_SETUP_PATH on hosts"
-    exit 1
-fi
 
-if [[ -n "$ydbd_tar" ]]; then
-    tar_name=$(basename "$ydbd_tar")
-    $debug parallel-scp -h "$HOSTS_FILE" -t 0 -p 20 "$ydbd_tar" "$YDB_SETUP_PATH"
     if [[ $? -ne 0 ]]; then
-        echo "ERROR: can't copy $ydbd_tar to hosts"
+        echo "ERROR: can't create $YDB_SETUP_PATH on hosts"
         exit 1
     fi
-elif [[ -n "$ydbd_url" ]]; then
-    tar_name=$(basename $ydbd_url)
-    $debug parallel-ssh -h "$HOSTS_FILE" -t 0 -p 20 "wget -q $ydbd_url -O $YDB_SETUP_PATH/$tar_name"
+
+    if [[ -n "$ydbd_tar" ]]; then
+        tar_name=$(basename "$ydbd_tar")
+        $debug parallel-scp -h "$HOSTS_FILE" -t 0 -p 20 "$ydbd_tar" "$YDB_SETUP_PATH"
+        if [[ $? -ne 0 ]]; then
+            echo "ERROR: can't copy $ydbd_tar to hosts"
+            exit 1
+        fi
+    elif [[ -n "$ydbd_url" ]]; then
+        tar_name=$(basename $ydbd_url)
+        $debug parallel-ssh -h "$HOSTS_FILE" -t 0 -p 20 "wget -q $ydbd_url -O $YDB_SETUP_PATH/$tar_name"
+        if [[ $? -ne 0 ]]; then
+            echo "ERROR: can't download $ydbd_url to hosts"
+            exit 1
+        fi
+    fi
+
+    $debug parallel-ssh -i -h "$HOSTS_FILE" -t 0 -p 20 "\
+        tar -xzf $YDB_SETUP_PATH/$tar_name --strip-component=1 -C $YDB_SETUP_PATH; \
+        rm -f $YDB_SETUP_PATH/$tar_name"
+
     if [[ $? -ne 0 ]]; then
-        echo "ERROR: can't download $ydbd_url to hosts"
+        echo "ERROR: can't extract $tar_name on hosts"
         exit 1
     fi
 else
-    echo "ERROR: ydbd tar or url is not specified"
-    usage
-    exit 1
-fi
-
-$debug parallel-ssh -i -h "$HOSTS_FILE" -t 0 -p 20 "\
-    tar -xzf $YDB_SETUP_PATH/$tar_name --strip-component=1 -C $YDB_SETUP_PATH; \
-    rm -f $YDB_SETUP_PATH/$tar_name"
-
-if [[ $? -ne 0 ]]; then
-    echo "ERROR: can't extract $tar_name on hosts"
-    exit 1
+    echo "Deploy skipped"
 fi
 
 $debug parallel-scp -h "$HOSTS_FILE" -t 0 -p 20 "$CONFIG_DIR"/config.yaml "$CONFIG_DIR"/config_dynnodes.yaml "$YDB_SETUP_PATH/cfg"
@@ -159,15 +155,17 @@ if [[ $? -ne 0 ]]; then
     exit 1
 fi
 
-log "Format disks"
+if [[ -z "$restart_ydb" ]]; then
+    log "Format disks"
 
-for d in "${DISKS[@]}"; do
-    $debug parallel-ssh -i -h "$HOSTS_FILE" -t 0 -p 20 "sudo LD_LIBRARY_PATH=$YDB_SETUP_PATH/lib $YDB_SETUP_PATH/bin/ydbd admin bs disk obliterate $d"
-    if [[ $? -ne 0 ]]; then
-        echo "ERROR: can't obliterate disk $d on hosts"
-        exit 1
-    fi
-done
+    for d in "${DISKS[@]}"; do
+        $debug parallel-ssh -i -h "$HOSTS_FILE" -t 0 -p 20 "sudo LD_LIBRARY_PATH=$YDB_SETUP_PATH/lib $YDB_SETUP_PATH/bin/ydbd admin bs disk obliterate $d"
+        if [[ $? -ne 0 ]]; then
+            echo "ERROR: can't obliterate disk $d on hosts"
+            exit 1
+        fi
+    done
+fi
 
 GRPC_PORT=$GRPC_PORT_BEGIN
 IC_PORT=$IC_PORT_BEGIN
@@ -197,24 +195,26 @@ for host in `cat $HOSTS_FILE`; do
     fi
 done
 
-log "Init BS"
+if [[ -z "$restart_ydb" ]]; then
+    log "Init BS"
 
-$debug ssh "$init_host" \
-    "sudo LD_LIBRARY_PATH=$YDB_SETUP_PATH/lib $YDB_SETUP_PATH/bin/ydbd admin blobstorage config init --yaml-file $YDB_SETUP_PATH/cfg/config.yaml"
+    $debug ssh "$init_host" \
+        "sudo LD_LIBRARY_PATH=$YDB_SETUP_PATH/lib $YDB_SETUP_PATH/bin/ydbd admin blobstorage config init --yaml-file $YDB_SETUP_PATH/cfg/config.yaml"
 
-if [[ $? -ne 0 ]]; then
-    echo "ERROR: can't init blobstorage on $init_host"
-    exit 1
-fi
+    if [[ $? -ne 0 ]]; then
+        echo "ERROR: can't init blobstorage on $init_host"
+        exit 1
+    fi
 
-log "Create storage pools"
+    log "Create storage pools"
 
-$debug ssh "$init_host" \
-    "sudo LD_LIBRARY_PATH=$YDB_SETUP_PATH/lib $YDB_SETUP_PATH/bin/ydbd admin database /Root/$DATABASE_NAME create $STORAGE_POOLS"
+    $debug ssh "$init_host" \
+        "sudo LD_LIBRARY_PATH=$YDB_SETUP_PATH/lib $YDB_SETUP_PATH/bin/ydbd admin database /Root/$DATABASE_NAME create $STORAGE_POOLS"
 
-if [[ $? -ne 0 ]]; then
-    echo "ERROR: can't create storage pools $STORAGE_POOLS on $init_host"
-    exit 1
+    if [[ $? -ne 0 ]]; then
+        echo "ERROR: can't create storage pools $STORAGE_POOLS on $init_host"
+        exit 1
+    fi
 fi
 
 for ind in $(seq 0 $(($DYNNODE_COUNT-1))); do
